Update README.md
It's arguable whether you need to take this. But I thought it was worth being clear that `debug` in this context refers to an npm package. I hadn't heard of it myself and first read this thinking that debug was referring to the activity of debugging. So I was confused